Removed unnecessary functions
